+++ php_swoole.h
@@ -454,7 +454,7 @@ void php_swoole_register_shutdown_function_prepend(char *function);
 void php_swoole_event_init();
 void php_swoole_event_wait();
 void php_swoole_event_exit();
-long php_swoole_add_timer(int ms, zval *callback, zval *param, int persistent);
+long php_swoole_add_timer(long ms, zval *callback, zval *param, int persistent);
 void php_swoole_register_callback(swServer *serv);
 void php_swoole_trace_check(void *arg);
 void php_swoole_client_free(zval *zobject, swClient *cli);
